% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{botox_dystonia}
\alias{botox_dystonia}
\title{\code{botox_dystonia}: Indomethicin vs. Placebo for Post-ERCP Pancreatitis (PEP)}
\format{
\subsection{\code{botox_dystonia}}{

A data frame with 545 rows and 9 columns:
\describe{
\item{participant_id}{Participant ID}
\item{site}{Site ID}
\item{age}{Age at Baseline}
\item{sex}{Sex}
\item{arm}{Study Arm}
\item{tx}{Treatment Indicator}
\item{twstrs_bl}{TWSTRS Total Score (BL)}
\item{week}{Week Post Randomization}
\item{twstrs}{TWSTRS Total Score - Followup}
}
}
}
\source{
\url{https://github.com/higgi13425/medicaldata}
}
\usage{
botox_dystonia
}
\description{
This is a randomized controlled trial conducted from 2009 to 2011 of 545
individuals, evaluating the efficacy of botulinium toxin B for cervical
dystonia (Brashear et al. 1999).
}
\details{
This data was originally sourced from the
\code{medicaldata} package (Higgins 2022), and modified to be ready to use.

Brashear, A., M. F. Lew, D. D. Dykstra, C. L. Comella, S. A. Factor,
R. L. Rodnitzky, R. Trosch, et al. 1999. "Safety and Efficacy of
NeuroBloc (Botulinum Toxin Type b) in Type a–Responsive Cervical
Dystonia." Neurology 53 (7): 1439–39.
https://doi.org/10.1212/wnl.53.7.1439.

Higgins, Peter. 2022. Medicaldata: Data Package for Medical Datasets.
https://github.com/higgi13425/medicaldata.
}
\keyword{datasets}
