% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon_cancer}
\alias{colon_cancer}
\title{\code{colon_cancer}: 5-Fluorouracil for Colon Cancer}
\format{
\subsection{\code{colon_cancer}}{

A data frame with 929 rows and 19 columns:
\describe{
\item{participant_id}{Participant ID}
\item{arm}{Study Arm}
\item{age_bl}{Age (BL)}
\item{sex}{Sex (BL)}
\item{obstruction_bl}{Tumor Obstructs Colon (BL)}
\item{perforation_bl}{Tumor Perforates Colon (BL)}
\item{organ_adherence_bl}{Tumor Organ Adherence (BL)}
\item{positive_nodes_bl}{Lymph Nodes with Cancer (BL)}
\item{differentiation_bl}{Tumor Differentiation (BL)}
\item{local_spread_bl}{Extent of Tumor Spread (BL)}
\item{time_surgery_registration_bl}{Time: Surgery to Registration (BL)}
\item{event_death}{Death On Study}
\item{time_to_death}{Time to Death On Study (d)}
\item{event_recurrence}{Recurrence On Study}
\item{time_to_recurrence}{Time to Recurrence On Study (d)}
\item{tx}{Treatment Indicator}
\item{years_to_death}{Time to Death On Study (Y)}
\item{years_to_recurrence}{Time to Recurrence On Study (Y)}
\item{cox_expected_obs}{Participant ID}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/survival/index.html}
}
\usage{
colon_cancer
}
\description{
This is a randomized controlled trial, enrolling between 1984 to 1987,
comparing survival and cancer recurrence among 929 participants
randomized to 5-Fluorouracil + levamisole, levamisole alone, or
observation (Laurie et al. 1989).
}
\details{
This data was originally sourced from the \code{survival} package
(Therneau 2026), and modified to be ready to use. Note that covariates
have been imputed using Multiple Imputation with Chained Equations
(MICE) using the \code{mice} package. This contains all experimental arms.

Laurie, J A, C G Moertel, T R Fleming, H S Wieand, J E Leigh, J Rubin,
G W McCormack, J B Gerstner, J E Krook, and J Malliard. 1989. "Surgical
Adjuvant Therapy of Large-Bowel Carcinoma: An Evaluation of Levamisole
and the Combination of Levamisole and Fluorouracil. The North Central
Cancer Treatment Group and the Mayo Clinic." Journal of Clinical
Oncology 7 (10): 1447–56. https://doi.org/10.1200/jco.1989.7.10.1447.

Therneau, Terry M. 2026. A Package for Survival Analysis in r.
https://CRAN.R-project.org/package=survival.
}
\keyword{datasets}
