% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indo_rct}
\alias{indo_rct}
\title{\code{indo_rct}: Indomethicin vs. Placebo for Post-ERCP Pancreatitis (PEP)}
\format{
\subsection{\code{indo_rct}}{

A data frame with 602 rows and 22 columns:
\describe{
\item{participant_id}{Participant ID}
\item{arm}{Study Arm}
\item{tx}{Treatment Indicator}
\item{age}{Age at Baseline}
\item{gender}{Gender}
\item{previous_pep_bl}{Previous PEP (BL)}
\item{pancreatitis_bl}{Recurrent Pancreatitis (BL)}
\item{p_sphincterotmy_bl}{Pancreatic Sphincterotomy (BL)}
\item{sphincter_precut_bl}{Sphincter Pre-cut Needed (BL)}
\item{difficult_cannulation_bl}{Difficult Cannulation of Papilla (BL)}
\item{pneumatic_dilation_bl}{Pneumatic Dilation of Papilla (BL)}
\item{ampullectomy_bl}{Ampullectomy Performed(BL)}
\item{injected_contrast_bl}{Contrast Injected (BL)}
\item{aspirin_any_bl}{Any Aspirin Use (BL)}
\item{any_p_duct_stent_bl}{Pancreatic Duct Stent Used (BL)}
\item{oddi_manometry_bl}{Sphicter of Oddi Manometry (BL)}
\item{b_sphincterotmy_bl}{Biliary Sphincterotomy (BL)}
\item{trainee_bl}{Trainee Involved in ERCP(BL)}
\item{is_outpatient_bl}{Outpatient Status (BL)}
\item{oddi_dysfunction_bl}{Sphincter of Oddi Dysfunction (BL)}
\item{pep}{Post-ERCP Pancreatitis}
\item{pep_binary}{Post-ERCP Pancreatitis}
}
}
}
\source{
\url{https://github.com/higgi13425/medicaldata}
}
\usage{
indo_rct
}
\description{
This is a randomized controlled trial conducted from 2009 to 2011 of 602
individuals, evaluating the efficacy of indomethicin for prevention of
Post-ERCP Pancreatitis (Elmunzer et al. 2012).
}
\details{
This data was originally sourced from the
\code{medicaldata} package (Higgins 2022), and modified to be ready to use.

Elmunzer, B. Joseph, James M. Scheiman, Glen A. Lehman, Amitabh Chak,
Patrick Mosler, Peter D. R. Higgins, Rodney A. Hayward, et al. 2012.
"A Randomized Trial of Rectal Indomethacin to Prevent Post-ERCP
Pancreatitis." New England Journal of Medicine 366 (15): 1414–22.
https://doi.org/10.1056/nejmoa1111103.

Higgins, Peter. 2022. Medicaldata: Data Package for Medical Datasets.
https://github.com/higgi13425/medicaldata.
}
\keyword{datasets}
