% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laryngoscope}
\alias{laryngoscope}
\title{\code{laryngoscope}: Macintosh vs. Pentax Laryngoscope for Orotracheal Intubation}
\format{
\subsection{\code{laryngoscope}}{

A data frame with 99 rows and 24 columns:
\describe{
\item{participant_id}{Participant ID}
\item{age_bl}{Age at Baseline}
\item{gender}{Gender}
\item{asa_physical_status_bl}{ASA Physical Status}
\item{mallampati_bl}{Mallampati Score (BL)}
\item{bmi_bl}{Body Mass Index (kg/m^2)}
\item{arm}{Study Arm}
\item{tx}{Treatment Indicator}
\item{attempt_1_time}{Attempt 1 Time (s)}
\item{attempt_1}{Attempt 1 Outcome}
\item{attempt_1_f}{Attempt 1 Outcome}
\item{attempt_2_time}{Attempt 2 Time (s)}
\item{attempt_2}{Attempt 2 Outcome}
\item{attempt_2_f}{Attempt 2 Outcome}
\item{attempt_3_time}{Attempt 3 Time (s)}
\item{attempt_3}{Attempt 3 Outcome}
\item{attempt_3_f}{Attempt 3 Outcome}
\item{attempts}{Intubation Attempts}
\item{failures}{Intubation Failures}
\item{total_intubation_time}{Total Intubation Time (s)}
\item{bleeding}{Post-Intubation Bleeding}
\item{difficulty}{Intubation Difficulty: Easy (0) - Difficult (100)}
\item{sore_throat}{Post-Intubation Sore Throat}
\item{glottic_view_grade}{Grade of Glottic View}
}
}
}
\source{
\url{https://github.com/higgi13425/medicaldata}
}
\usage{
laryngoscope
}
\description{
This is a randomized controlled trial of 99 individuals, evaluating the
relative efficacy of two laryngoscope models for orotracheal intubation
in obese patients (Abdallah et al. 2011).
}
\details{
This data was originally sourced from the
\code{medicaldata} package (Higgins 2022), and modified to be ready to use.

Abdallah, Rania, Ursula Galway, Jing You, Andrea Kurz, Daniel I.
Sessler, and D. John Doyle. 2011. "A Randomized Comparison Between the
Pentax AWS Video Laryngoscope and the Macintosh Laryngoscope in Morbidly
Obese Patients." Anesthesia & Analgesia 113 (5): 1082–87.
https://doi.org/10.1213/ane.0b013e31822cf47d.

Higgins, Peter. 2022. Medicaldata: Data Package for Medical Datasets.
https://github.com/higgi13425/medicaldata.
}
\keyword{datasets}
