% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{licorice_gargle}
\alias{licorice_gargle}
\title{\code{licorice_gargle}: Licorice Gargle vs. Sucrose for Management of
Endotracheal Intubation}
\format{
\subsection{\code{licorice_gargle}}{

A data frame with 235 rows and 25 columns:
\describe{
\item{participant_id}{Participant ID}
\item{age_bl}{Age at Baseline}
\item{gender}{Gender}
\item{asa_physical_status_bl}{ASA Physical Status}
\item{bmi_bl}{Body Mass Index (kg/m^2)}
\item{mallampati_bl}{Mallampati Score (BL)}
\item{smoking_bl}{Smoking Status (BL)}
\item{pain_yn_bl}{Preoperative Pain (BL)}
\item{arm}{Study Arm}
\item{tx}{Treatment Indicator}
\item{surgery_size_post_rnd}{Surgery Size (Post-Randomization)}
\item{extubation_cough}{Cough At Extubation}
\item{pacu_30_min_cough}{Cough at PACU: 30 Minutes}
\item{pacu_30_min_throat_pain}{Throat Pain at PACU: 30 Minutes}
\item{pacu_30_min_any_throat_pain}{Any Throat Pain at PACU: 30 Minutes}
\item{pacu_30_min_swallow_pain}{Swallow Pain at PACU: 30 Minutes}
\item{pacu_90_min_cough}{Cough at PACU: 90 Minutes}
\item{pacu_90_min_throat_pain}{Throat Pain at PACU: 90 Minutes}
\item{pacu_90_min_any_throat_pain}{Any Throat Pain at PACU: 90 Minutes}
\item{postop_4h_cough}{Cough: 4H Post-Op}
\item{postop_4h_throat_pain}{Throat Pain: 4H Post-Op}
\item{postop_4h_any_throat_pain}{Any Throat Pain: 4H Post-Op}
\item{postop_1d_am_cough}{Cough: 1d Post-Op in AM}
\item{postop_1d_am_throat_pain}{Throat Pain: 1d Post-Op in AM}
\item{postop_1d_am_any_throat_pain}{Any Throat Pain: 1d Post-Op in AM}
}
}
}
\source{
\url{https://github.com/higgi13425/medicaldata}
}
\usage{
licorice_gargle
}
\description{
This is a randomized controlled trial, conducted from 2010 to 2011,
evaluating whether a licorice or sucrose gargle provided better
management of endotracheal tube extubation symptoms in 235 participants
(Ruetzler et al. 2013).
}
\details{
This data was originally sourced from the
\code{medicaldata} package (Higgins 2022), and modified to be ready to use.

Ruetzler, Kurt, Michael Fleck, Sabine Nabecker, Kristina Pinter, Gordian
Landskron, Andrea Lassnigg, Jing You, and Daniel I. Sessler. 2013. "A
Randomized, Double-Blind Comparison of Licorice Versus Sugar-Water
Gargle for Prevention of Postoperative Sore Throat and Postextubation
Coughing." Anesthesia & Analgesia 117 (3): 614–21.
https://doi.org/10.1213/ane.0b013e318299a650.

Higgins, Peter. 2022. Medicaldata: Data Package for Medical Datasets.
https://github.com/higgi13425/medicaldata.
}
\keyword{datasets}
