% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung_cancer}
\alias{lung_cancer}
\title{\code{lung_cancer}: Veteran’s Administration Lung Cancer Trial}
\format{
\subsection{\code{lung_cancer}}{

A data frame with 137 rows and 9 columns:
\describe{
\item{participant_id}{Participant ID}
\item{age_bl}{Age (BL)}
\item{cell_type_bl}{Tumor Cell Type (BL)}
\item{karnofsky_bl}{Karnofsky Performance Score (BL)}
\item{time_from_dx_bl}{Time: Diagnosis to Randomization (BL)}
\item{arm}{Study Arm}
\item{tx}{Treatment Indicator}
\item{time_to_event}{Time to Death On Study (d)}
\item{event}{Death On Study}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/survival/index.html}
}
\usage{
lung_cancer
}
\description{
This is a randomized controlled trial conducted by the Veteran’s
Administration Lung Cancer group, comparing survival among 137
participants randomized to two treatment regimens.
}
\details{
This data was originally sourced from the \code{survival} package
(Therneau 2026), and modified to be ready to use.

Prentice, R. L. 1973. "Exponential Survivals with Censoring and
Explanatory Variables." Biometrika 60 (2): 279–88.
https://doi.org/10.1093/biomet/60.2.279.

Kalbfleisch, John D., and Ross L. Prentice. 2002. The Statistical
Analysis of Failure Time Data. Wiley Series in Probability and
Statistics. Wiley. https://doi.org/10.1002/9781118032985.

Therneau, Terry M. 2026. A Package for Survival Analysis in r.
https://CRAN.R-project.org/package=survival.
}
\keyword{datasets}
