% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perio_pregnancy}
\alias{perio_pregnancy}
\title{\code{perio_pregnancy}: Maternal Periodontal Therapy and Pregnancy Outcomes}
\format{
\subsection{\code{perio_pregnancy}}{

A data frame with 823 rows and 23 columns:
\describe{
\item{participant_id}{Participant ID}
\item{age_bl}{Mother's Age (BL)}
\item{education_bl}{Education (BL)}
\item{public_assistance_bl}{On Public Assistance (BL)}
\item{hypertension_bl}{Hypertension (BL)}
\item{diabetes_bl}{Any Diabetes (BL)}
\item{diabetes_type_bl}{Diabetes Type (BL)}
\item{bmi_bl}{Body Mass Index - kg/m^2 (BL)}
\item{tobacco_use_bl}{Tobacco Use (BL)}
\item{previous_pregnancy_bl}{Previous Pregnancy (BL)}
\item{previous_live_preterm_bl}{Previous Live Preterm Birth (BL)}
\item{previous_stillbirth_bl}{Previous Stillbirth (BL)}
\item{previous_spontaneous_abortion}{Previous Spontaneous Abortion (BL)}
\item{previous_induced_abortion}{Previous Induced Abortion (BL)}
\item{any_lptb_sb_sa_ia_bl}{Previous LPTB, SB, SA, or IA (BL)}
\item{site}{Study Site}
\item{arm}{Study Arm}
\item{tx}{Treatment Indicator}
\item{birth_outcome}{Birth Outcome}
\item{birthweight}{Birthweight at Outcome}
\item{gestational_age}{Gestational Age at Outcome}
\item{apgar_1_min}{Apgar (1m)}
\item{apgar_5_min}{Apgar (5m)}
}
}
}
\source{
\url{https://github.com/higgi13425/medicaldata}
}
\usage{
perio_pregnancy
}
\description{
This is a randomized controlled trial, conducted from 1984 to 2005,
evaluating whether maternal periodontic treatment 823 participants
(Michalowicz et al. 2006).
}
\details{
This data was originally sourced from the
\code{medicaldata} package (Higgins 2022), and modified to be ready to use.

Michalowicz, Bryan S., James S. Hodges, Anthony J. DiAngelis, Virginia
R. Lupo, M. John Novak, James E. Ferguson, William Buchanan, et al.
2006. "Treatment of Periodontal Disease and the Risk of Preterm Birth."
New England Journal of Medicine 355 (18): 1885–94.
https://doi.org/10.1056/nejmoa062249.

Higgins, Peter. 2022. Medicaldata: Data Package for Medical Datasets.
https://github.com/higgi13425/medicaldata.
}
\keyword{datasets}
