% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strep_tb}
\alias{strep_tb}
\title{\code{strep_tb}: Streptomycin for Tuberculosis}
\format{
\subsection{\code{strep_tb}}{

A data frame with 107 rows and 13 columns:
\describe{
\item{patient_id}{Participant ID}
\item{arm}{Study Arm}
\item{tx}{Treatment Indicator}
\item{gender}{Gender}
\item{condition_bl}{Condition (BL)}
\item{temperature_bl}{Temperature  (BL)}
\item{sed_rate_esr_bl}{ESR: Sedimentation Rate (BL)}
\item{cxr_lung_cavitation_bl}{Lung Cavitation on X-Ray (BL)}
\item{strep_resistance_6m}{Strep Resistance (6M)}
\item{radiologic_outcome_6m}{Radiologic Outcome (6M)}
\item{chest_xray_rating_6m}{Chest X-Ray Numeric Rating (6M)}
\item{radiologic_improvement_6m}{Radiologic Improvement (6M)}
\item{radiologic_improvement_f_6m}{Radiologic Improvement (6M)}
}
}
}
\source{
\url{https://github.com/higgi13425/medicaldata}
}
\usage{
strep_tb
}
\description{
This is a randomized controlled trial of 107 individuals, evaluating the
efficacy of streptomycin for pulmonary tuberculosis, conducted in 1947
(Medical Research Council 1998). Participants were randomized to receive
streptomycin and bed rest (treatment) or bed rest alone (control), and
followed up after six months.
}
\details{
This data was originally sourced from the
\code{medicaldata} package (Higgins 2022), and modified to be ready to use.

Medical Research Council. 1998. "Streptomycin Treatment of Pulmonary
Tuberculosis." Edited by British Medical Journal Publishing Group.
BMJ 317 (7167): 1248. https://doi.org/10.1136/bmj.317.7167.1248b.

Higgins, Peter. 2022. Medicaldata: Data Package for Medical Datasets.
https://github.com/higgi13425/medicaldata.
}
\keyword{datasets}
