% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supraclavicular}
\alias{supraclavicular}
\title{\code{supraclavicular}: Combined vs. Sequential Administration of
Supraclavicular Anaesthesia}
\format{
\subsection{\code{supraclavicular}}{

A data frame with 103 rows and 20 columns:
\describe{
\item{participant_id}{Participant ID}
\item{age_bl}{Age at Baseline (BL)}
\item{gender}{Gender (BL)}
\item{bmi_bl}{Body Mass Index - kg/m^2 (BL)}
\item{arm}{Study Arm}
\item{tx}{Treatment Indicator}
\item{fentanyl_intraop}{Intraoperative Fentanyl (ug)}
\item{alfentanil_intraop}{Intraoperative Alfentanyl (mg)}
\item{midazolam_intraop}{Intraoperative Midazolam (mg)}
\item{time_4_nerve_block}{Time: 4-Nerve Sensory Block (m)}
\item{time_1st_nerve_block}{Time: 1st Sensory Block (m)}
\item{nerve_block_achieved}{Achieve Sensory+Motor Block}
\item{nerve_block_achieved_f}{Achieve Sensory+Motor Block}
\item{time_complete_motor_block}{Time: Complete Motor Block (m)}
\item{time_block_to_meds}{Time: Nerve Block to Meds Requested (h)}
\item{meds_requested}{Analgesia Requested 2d Post-op}
\item{meds_requested_f}{Analgesia Requested 2d Post-op}
\item{max_postop_pain_rest}{Max Postop Verbal Pain (0-10): Rest}
\item{max_postop_pain_movement}{Max Postop Verbal Pain (0-10): Movement}
\item{total_opioids}{Total Opioids (mg)}
}
}
}
\source{
\url{https://github.com/higgi13425/medicaldata}
}
\usage{
supraclavicular
}
\description{
This is a randomized controlled trial conducted from 2009 to 2011 of 545
individuals, evaluating the efficacy of botulinium toxin B for cervical
dystonia (Brashear et al. 1999).
}
\details{
This data was originally sourced from the
\code{medicaldata} package (Higgins 2022), and modified to be ready to use.

Roberman, Dmitry, Harendra Arora, Daniel I. Sessler, Michael Ritchey,
Jing You, and Priya Kumar. 2011. "Combined Versus Sequential Injection
of Mepivacaine and Ropivacaine for Supraclavicular Nerve Blocks."
Regional Anesthesia and Pain Medicine 36 (2): 145–50.
https://doi.org/10.1097/aap.0b013e31820d4235.

Higgins, Peter. 2022. Medicaldata: Data Package for Medical Datasets.
https://github.com/higgi13425/medicaldata.
}
\keyword{datasets}
